LayoutTests/accessibility/element-reflection-ariaflowto-expected.txt

Checks that element reflection is exposed to the a11y tree for 'ariaFlowToElements'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axTarget1.ariaFlowToElementAtIndex(0).isEqual(axFlow1) is true
PASS axTarget2.ariaFlowToElementAtIndex(0).isEqual(axFlow2) is true
PASS axTarget2.ariaFlowToElementAtIndex(0).isEqual(axFlow1) is true
PASS axInnerTarget.ariaFlowToElementAtIndex(0).isEqual(axFlow3) is true
PASS axTarget2.ariaFlowToElementAtIndex(0).isEqual(axFlow1) is true
PASS axTarget2.ariaFlowToElementAtIndex(1).isEqual(axFlow2) is true
PASS axTarget2.ariaFlowToElementAtIndex(2).isEqual(axFlow3) is true
PASS axTarget4.ariaFlowToElementAtIndex(0).isEqual(axFlow4) is true
PASS axTarget5.ariaFlowToElementAtIndex(0).isEqual(axFlow5) is true
PASS successfullyParsed is true

TEST COMPLETE
First flow
Target 1
Second flow
Target 2
Third flow
Fourth flow
Target 4
Fifth flow
